Factor out nb connector
