+++ connectors/dynamo/__init__.py
@@ -1,1 +1,1 @@
-__all__ = ["DomainModel", "StoryModel", "DynamoDbClient"]
+__all__ = ["DomainModel", "StoryModel", "DynamoDbClient", "ErrorModel"]
