+++ exceptions/BlacklistedError.py
@@ -1,0 +1,3 @@
+class BlacklistedError(Exception):
+    def __init__(self):
+        pass
