+++ requirements.txt
@@ -5,3 +5,4 @@ wsgiref==0.1.2
 MySQL-python==1.2.5
 pynamodb==3.3.0
 datadog==0.21.0
+ddtrace==0.12.1
