+++ utility/__init__.py
@@ -1,1 +1,1 @@
-__all__ = ["constants", "nb_logging"]
+__all__ = ["constants", "nb_logging", "client_factory"]
